+++ CHANGELOG.md
@@ -14,6 +14,7 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 - Ensured installation of `os2web_audit`
   - `os2web_audit` has default logging through `watchdog`
 - Adds audit logging to `os2forms_digital_post`
+- Adds audit logging to `os2forms_nemid`
 
 ## [3.17.0] 2024-11-21